==> 3. Final/05-4) TF-IDF & Logistic(연도별 CSV).ipynb <==
# **TF-IDF & Logistic**

import numpy as np
import pandas as pd

# Visualization
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

# DB 연결
# !pip install pymysql
import pymysql

# DB에 저장
import sqlalchemy
from sqlalchemy import create_engine

# 실시간 주식가격 데이터
# !pip install finance-datareader
import FinanceDataReader as fdr

# 텍스트 분석
# 1) JAVA 설치, 2) Python 버전과 맞는 JPype1-py3 설치, 3) !pip install konlpy, 4) 설치 경로에서 jvm.py 파일 코드 67번 줄 주석 처리 
from konlpy.tag import Okt
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import re

# 모델 학습 및 평가
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.metrics import precision_score, recall_score, f1_score

# 모델 저장 및 로드
import joblib

# 기타
import datetime
from collections import Counter
import sys
import warnings
warnings.filterwarnings('ignore')

## **Read & Save Data**

# 매일경제
maeil_2018 = pd.read_csv('../../../../Code/Data/Test/2018_maeil.csv')
maeil_2019 = pd.read_csv('../../../../Code/Data/Test/2019_maeil.csv')
maeil_2020 = pd.read_csv('../../../../Code/Data/Test/2020_maeil.csv')
maeil_2021 = pd.read_csv('../../../../Code/Data/Test/2021_maeil.csv')
# 아시아경제
asia_2018 = pd.read_csv('../../../../Code/Data/Test/2018_asia.csv')
asia_2019 = pd.read_csv('../../../../Code/Data/Test/2019_asia.csv')
asia_2020 = pd.read_csv('../../../../Code/Data/Test/2020_asia.csv')
asia_2021 = pd.read_csv('../../../../Code/Data/Test/2021_asia.csv')

# 연도별 통합
news_2018 = pd.concat([maeil_2018, asia_2018], axis=0)
news_2019 = pd.concat([maeil_2019, asia_2019], axis=0)
news_2020 = pd.concat([maeil_2020, asia_2020], axis=0)
news_2021 = pd.concat([maeil_2021, asia_2021], axis=0)

# 연도별 TF-IDF & Logistic 적용을 위해 CSV 파일로 저장
news_2018.to_csv('../../../../Code/Data/Test/news_2018.csv', index=False)
news_2019.to_csv('../../../../Code/Data/Test/news_2019.csv', index=False)
news_2020.to_csv('../../../../Code/Data/Test/news_2020.csv', index=False)
news_2021.to_csv('../../../../Code/Data/Test/news_2021.csv', index=False)

# 2018~2020년 데이터로 Train-Validation Set 구성
news_18to20 = pd.concat([maeil_2018, asia_2018, maeil_2019, asia_2019, maeil_2020, asia_2020], axis=0)
news_18to20.to_csv('../../../../Code/Data/Test/news_18to20.csv', index=False)
