==> 3. Final/06-1) Lexicon(All Data).ipynb <==
# **Lexicon**

import numpy as np
import pandas as pd

### Visualization
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### DB 연결
# !pip install pymysql
import pymysql

### DB에 저장
import sqlalchemy
from sqlalchemy import create_engine

### 실시간 주식가격 데이터
# !pip install finance-datareader
import FinanceDataReader as fdr

### 텍스트 분석
## KoNLPy
# 1) JAVA 설치, 2) Python 버전과 맞는 JPype1-py3 설치, 3) !pip install konlpy, 4) 설치 경로에서 jvm.py 파일 코드 67번 줄 주석 처리 
from konlpy.tag import Okt
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import re
## FastText
# !pip install gensim
# !pip install fasttext
import fasttext
import fasttext.util
# Facebook 한국어 Embedding 모델 다운로드 → 한 번만 설치하면 됨
fasttext.util.download_model('ko', if_exists='ignore')
# 유사도 계산
from gensim import models

### 모델 학습 및 평가
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.metrics import precision_score, recall_score, f1_score

### 모델 저장 및 로드
import joblib

### 기타
import datetime
from collections import Counter
import sys
import warnings
warnings.filterwarnings('ignore')

import tensorflow as tf

print('GPU', '사용 가능' if tf.config.experimental.list_physical_devices('GPU') else '사용 불가능')

!nvidia-smi

## **Read KOSELF & Test Data**

### **① KOSELF 감성 어휘 사전**

# 블로그에서 가져온 기본적인 한국어 긍부정 텍스트 목록
with open('positive_words_self.txt', encoding='utf-8') as pos_blog:
    positive_blog = pos_blog.readlines()
positive_blog = [pos_blog.replace('\n', '') for pos_blog in positive_blog]
with open('negative_words_self.txt', encoding='utf-8') as neg_blog:
    negative_blog = neg_blog.readlines()
negative_blog = [neg_blog.replace('\n', '') for neg_blog in negative_blog]

# KOSELF 감성 어휘 사전
with open('KOSELF_pos.txt', encoding='utf-8') as pos:
    positive = pos.readlines()
positive = [pos.replace('\n', '') for pos in positive]
with open('KOSELF_neg.txt', encoding='utf-8') as neg:
    negative = neg.readlines()
negative = [neg.replace('\n', '') for neg in negative]

### **② 2018~2020년 Data**

news_18to20 = pd.read_csv('../../../../Code/Data/Test/news_18to20.csv')

print(news_18to20.shape)
news_18to20.head(1)

news_18to20.dropna(inplace=True)
news_18to20 = news_18to20.reset_index(drop=True)

## **Calculate Cosine Similarity**

# 이미 학습된 FastText 내장 한국어 모델
try:
    print(ko_model, '모델 로드 불필요')
except:
    ko_model = models.fasttext.load_facebook_model('cc.ko.300.bin')

# KOSELF_neg와의 Cosine Similarity 계산
pos_cosine = []
for x in range(len(news_18to20['except_stopwords'])):
    for y in range(len(positive)):
        for z in range(len(news_18to20['except_stopwords'][x].split())):
            if ko_model.wv.similarity(positive[y], news_18to20['except_stopwords'][x].split()[z]) > 0.5:
                pos_cosine.append(news_18to20['except_stopwords'][x].split()[z])

pos_cosine = list(set(pos_cosine))

pos_cosine

len(pos_cosine)

# KOSELF_neg와의 Cosine Similarity 계산
neg_cosine = []
for x in range(len(news_18to20['except_stopwords'])):
    for y in range(len(negative)):
        for z in range(len(news_18to20['except_stopwords'][x].split())):
            if ko_model.wv.similarity(negative[y], news_18to20['except_stopwords'][x].split()[z]) > 0.5:
                pos_cosine.append(news_18to20['except_stopwords'][x].split()[z])

pos_cosine = list(set(neg_cosine))

neg_cosine

## **Read Data**

### **① KOSELF 감성 어휘 사전**

# 블로그에서 가져온 기본적인 한국어 긍부정 텍스트 목록
with open('positive_words_self.txt', encoding='utf-8') as pos_blog:
    positive_blog = pos_blog.readlines()
positive_blog = [pos_blog.replace('\n', '') for pos_blog in positive_blog]
with open('negative_words_self.txt', encoding='utf-8') as neg_blog:
    negative_blog = neg_blog.readlines()
negative_blog = [neg_blog.replace('\n', '') for neg_blog in negative_blog]

# KOSELF 감성 어휘 사전
with open('KOSELF_pos.txt', encoding='utf-8') as pos:
    positive = pos.readlines()
positive = [pos.replace('\n', '') for pos in positive]
with open('KOSELF_neg.txt', encoding='utf-8') as neg:
    negative = neg.readlines()
negative = [neg.replace('\n', '') for neg in negative]

### **② 연도별 News Data**

years = ['2018', '2019', '2020', '2021']
for i in range(len(years)):
    globals()['news_' + years[i]] = pd.read_csv('../../../../Code/Data/news_{}.csv'.format(years[i]))
    globals()['news_' + years[i]].dropna(axis=0, inplace=True)

test = pd.concat([news_2018, news_2019, news_2020])

test.shape

test[test['text'].isna()]

## **③ Stop Words**

# https://gist.github.com/spikeekips/40eea22ef4a89f629abd87eed535ac6a#file-stopwords-ko-txt
with open('stopwords-ko.txt', encoding='utf-8') as sw:
    stop_words = sw.readlines()
stop_words = [sw.replace('\n', '') for sw in stop_words]

## **Sentiment Analysis**

### **① 연도별 Lexicon**

#### **(1) Pos Dict**

tokenizer = Okt()

for i in range(len(years)):
    globals()['pos_' + years[i]] = []
    positive_text = globals()['news_' + years[i]][globals()['news_' + years[i]]['Extremely_Changed']==1]['text']
    
    for x in positive_text:
        words = tokenizer.nouns(x)
        # 불용어 제거하고 긍정어 리스트에 추가
        for y in words:
            if y not in stop_words:
                globals()['pos_' + years[i]].append(y)
            else:
                pass

#### **(2) Neg Dict**

tokenizer = Okt()

for i in range(len(years)):
    globals()['neg_' + years[i]] = []
    negative_text = globals()['news_' + years[i]][globals()['news_' + years[i]]['Extremely_Changed']==1]['text']
    
    for x in negative_text:
        words = tokenizer.nouns(x)
        # 불용어 제거하고 부정어 리스트에 추가
        for y in words:
            if y not in stop_words:
                globals()['neg_' + years[i]].append(y)
            else:
                pass

#### **(3) Giro Dict**

Counter(pos_2018).most_common(20)[0][1]

len(Counter(pos_2018)), len(pos_2018)

# 긍정어
for i in range(len(years)):
    globals()['lexicon_pos_' + years[i]] = pd.DataFrame(Counter(globals()['pos_' + years[i]]).most_common(), columns=['word', 'frequency'])

# 부정어
for i in range(len(years)):
    globals()['lexicon_neg_' + years[i]] = pd.DataFrame(Counter(globals()['neg_' + years[i]]).most_common(), columns=['word', 'frequency'])

lexicon_pos_2018
