==> 3. Final/09-2) Sentiment Analysis(Lexicon).ipynb <==
# **Sentiment Analysis(Lexicon)**

import numpy as np
import pandas as pd

### Visualization
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

### DB 연결
# !pip install pymysql
import pymysql

### DB에 저장
import sqlalchemy
from sqlalchemy import create_engine

### 실시간 주식가격 데이터
# !pip install finance-datareader
import FinanceDataReader as fdr

### 텍스트 분석
## KoNLPy
# 1) JAVA 설치, 2) Python 버전과 맞는 JPype1-py3 설치, 3) !pip install konlpy, 4) 설치 경로에서 jvm.py 파일 코드 67번 줄 주석 처리 
from konlpy.tag import Okt
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import re
## FastText
# !pip install gensim
# !pip install fasttext
import fasttext
import fasttext.util
# Facebook 한국어 Embedding 모델 다운로드 → 한 번만 설치하면 됨
# fasttext.util.download_model('ko', if_exists='ignore')   # FastText 모델 사용 시에만 필요
# 유사도 계산
from gensim import models

### 모델 학습 및 평가
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.metrics import precision_score, recall_score, f1_score

### 모델 저장 및 로드
import joblib

### 기타
import datetime
from collections import Counter
import sys
import warnings
warnings.filterwarnings('ignore')

## **Read Data**

### **① KOSELF + Blog 감성사전**

# 블로그에서 가져온 기본적인 한국어 긍부정 텍스트 목록
with open('positive_words_self.txt', encoding='utf-8') as pos_blog:
    positive_blog = pos_blog.readlines()
positive_blog = [pos_blog.replace('\n', '') for pos_blog in positive_blog]
with open('negative_words_self.txt', encoding='utf-8') as neg_blog:
    negative_blog = neg_blog.readlines()
negative_blog = [neg_blog.replace('\n', '') for neg_blog in negative_blog]

# KOSELF 감성 어휘 사전
with open('KOSELF_pos.txt', encoding='utf-8') as pos:
    positive = pos.readlines()
positive = [pos.replace('\n', '') for pos in positive]
with open('KOSELF_neg.txt', encoding='utf-8') as neg:
    negative = neg.readlines()
negative = [neg.replace('\n', '') for neg in negative]

len(positive), len(positive_blog), len(negative), len(negative_blog)

# 긍정사전 통합
for i in range(len(positive_blog)):
    if positive_blog[i] not in positive:
        positive.append(positive_blog[i])

# 부정사전 통합
for i in range(len(negative_blog)):
    if negative_blog[i] not in negative:
        negative.append(negative_blog[i])

len(positive), len(negative)

# 중복 확인
cnt = 0
for i in range(len(positive)):
    if positive[i] in negative:
        cnt += 1

cnt

### **② Text Data**

# 종목별, 연도별 데이터 통합
corp_list = ['samsung', 'hyundai', 'lg', 'sk', 'celltrion']
years = ['2018', '2019', '2020', '2021']
media_list = ['maeil', 'asia']

for i in range(len(corp_list)):
    globals()[corp_list[i] + '_news'] = []
    for j in range(len(years)):
        for k in range(len(media_list)):
            globals()[corp_list[i] + '_' + media_list[k] + '_' + years[j]] = pd.read_csv('../../../../Final Data/{0}_{1}_{2}.csv'.format(years[j], media_list[k], corp_list[i]))
            globals()[corp_list[i] + '_news'].append(globals()[corp_list[i] + '_' + media_list[k] + '_' + years[j]])

for i in range(len(corp_list)):
    globals()[corp_list[i]] = pd.concat(globals()[corp_list[i] + '_news'], axis=0)

## **Sentiment Score**

samsung[samsung['Tokenization'].isna()]

hyundai[hyundai['Tokenization'].isna()]

lg[lg['Tokenization'].isna()]

sk[sk['Tokenization'].isna()]

celltrion[celltrion['Tokenization'].isna()]

# 결측치 제거
for i in range(len(corp_list)):
    globals()[corp_list[i]].dropna(inplace=True)
    globals()[corp_list[i]] = globals()[corp_list[i]].reset_index(drop=True)

# iteration limit 조정
limit_number = 15000
sys.setrecursionlimit(limit_number)

for i in range(len(corp_list)):
    globals()[corp_list[i]]['Positive_Score'] = 0
    globals()[corp_list[i]]['Negative_Score'] = 0
    globals()[corp_list[i]]['Ratio'] = 0.1
    globals()[corp_list[i]]['Pred'] = 0
    globals()[corp_list[i]]['NSI'] = 0.1
    
    for j in range(len(globals()[corp_list[i]])):
        pos_score = 0 ; neg_score = 0
        
        for k in range(len(globals()[corp_list[i]]['Tokenization'][j].split())):
            if globals()[corp_list[i]]['Tokenization'][j].split()[k] in positive:
                pos_score += 1
            elif globals()[corp_list[i]]['Tokenization'][j].split()[k] in negative:
                neg_score += 1
            else:
                pass
        
        globals()[corp_list[i]]['Positive_Score'][j] = pos_score
        globals()[corp_list[i]]['Negative_Score'][j] = neg_score
        
        # 긍정과 부정의 비율
        if (pos_score==0) and (neg_score==0):
            globals()[corp_list[i]]['Ratio'][j] = 0.5   # 둘 다 0일 경우에는 긍정으로 가정
        else:
            globals()[corp_list[i]]['Ratio'][j] = pos_score / (pos_score + neg_score)
        
        # 예측 결과
        if globals()[corp_list[i]]['Ratio'][j]>=0.5:
            globals()[corp_list[i]]['Pred'][j] = 1
        else:
            globals()[corp_list[i]]['Pred'][j] = -1
        
        # 뉴스심리지수(NSI) 계산
        if (pos_score==0) and (neg_score==0):
            globals()[corp_list[i]]['NSI'][j] = 101
        else:
            globals()[corp_list[i]]['NSI'][j] = (pos_score - neg_score) / (pos_score + neg_score) * 100 + 100
    
    # CSV로 저장
    globals()[corp_list[i]].to_csv('../../../../Final Data/{}. {}_score.csv'.format(i+1, corp_list[i]), index=False)
    
    corp = corp_list[i].upper()
    print("... {}'s Sentiment Score 계산 완료! ...".format(corp))

print('       ########## 전부 완료-! ##########       ')

# Accuracy 확인
for i in range(len(corp_list)):
    corp = corp_list[i].upper()
    print('<{}> Acccuracy of {} : {}'.format(i+1, corp, (len(globals()[corp_list[i]][(globals()[corp_list[i]]['UpDown']<0) & (globals()[corp_list[i]]['Pred']<0)]) + len(globals()[corp_list[i]][(globals()[corp_list[i]]['UpDown']>0) & (globals()[corp_list[i]]['Pred']>0)])) / len(globals()[corp_list[i]])))
